fixed date format in user_detail tests, added more dummy data

+++ api/management/commands/populatedb.py
@@ -9,6 +9,8 @@ from api.data.db_dummy_data import (
     EXERCISES_USER_2,
     ROUTINES_USER_1,
     ROUTINES_USER_2,
+    WORKOUTS_USER_1,
+    WORKOUTS_USER_2,
 )
 from api.models import (
     Exercise,
@@ -22,7 +24,6 @@ from api.models import (
 )
 from django.contrib.auth.models import User
 from django.core.management.base import BaseCommand
-from lorem_text import lorem
 
 random.seed(0)
 
@@ -31,7 +32,11 @@ def create_exercise(
     name, kind, owner, instructions="", forks_count=0, tags=[], tutorials=[], muscles=[]
 ):
     exercise = Exercise(
-        name=name, kind=kind, instructions=instructions, owner=owner, forks_count=forks_count
+        name=name,
+        kind=kind,
+        instructions=instructions,
+        owner=owner,
+        forks_count=forks_count,
     )
     exercise.save()
 
@@ -45,25 +50,47 @@ def create_exercise(
 
 def create_routine(name, kind, owner, instructions="", forks_count=0, exercises=[]):
     routine = Routine.objects.create(
-        name=name, kind=kind, instructions=instructions, owner=owner, forks_count=forks_count
+        name=name,
+        kind=kind,
+        instructions=instructions,
+        owner=owner,
+        forks_count=forks_count,
     )
     for exercise_dict in exercises:
         exercise = Exercise.objects.get(name=exercise_dict.pop("name"))
         RoutineUnit.objects.create(exercise=exercise, routine=routine, **exercise_dict)
 
 
+def create_workout(date, owner, completed=False, routine=None, exercises=[]):
+    """
+
+    routine (str):
+        Should be passed as a routine name string.
+
+    """
+    if routine:
+        routine = Routine.objects.get(owner=owner, name=routine)
+
+    workout = Workout.objects.create(date=date, owner=owner, completed=completed, routine=routine)
+
+    print(workout)
+
+
 class Command(BaseCommand):
     help = "Command populating database"
 
     def reset_db(self):
+        apps = ["accounts", "api"]
         # Remove all migrations and database file
-        path_migrations = "api/migrations"
-        for file in os.listdir(path_migrations):
-            if file != "__init__.py":
-                try:
-                    os.remove(os.path.join(path_migrations, file))
-                except IsADirectoryError:
-                    shutil.rmtree(os.path.join(path_migrations, file))
+        for app in apps:
+            path_migrations = os.path.join(app, "migrations")
+            if os.path.exists(path_migrations):
+                for file in os.listdir(path_migrations):
+                    if file != "__init__.py":
+                        try:
+                            os.remove(os.path.join(path_migrations, file))
+                        except IsADirectoryError:
+                            shutil.rmtree(os.path.join(path_migrations, file))
 
         path_db = "db.sqlite3"
         try:
@@ -72,8 +99,10 @@ class Command(BaseCommand):
             pass
 
         # Recreate all tables
-        os.system("./manage.py makemigrations")
-        os.system("./manage.py migrate")
+        for app in apps:
+            os.system(f"./manage.py makemigrations {app}")
+            os.system(f"./manage.py migrate {app}")
+        os.system(f"./manage.py migrate")
 
     def create_users(self):
         # Create superuser (pk = 1)
@@ -85,7 +114,11 @@ class Command(BaseCommand):
 
         # First test user (pk=2)
         u1 = User.objects.create_user(
-            "test", "test@mail.com", password="test", first_name="Test", last_name="Smith"
+            "test",
+            "test@mail.com",
+            password="test",
+            first_name="Test",
+            last_name="Smith",
         )
         u1.profile.date_of_birth = datetime.date(1990, 1, 10)
         u1.profile.country = "Poland"
@@ -95,7 +128,11 @@ class Command(BaseCommand):
 
         # Second test user (pk=3)
         u2 = User.objects.create_user(
-            "test2", "test2@mail.com", password="test", first_name="Diana", last_name="Davis"
+            "test2",
+            "test2@mail.com",
+            password="test",
+            first_name="Diana",
+            last_name="Davis",
         )
         u2.profile.city = "New York"
         u2.profile.gender = "f"
@@ -120,12 +157,12 @@ class Command(BaseCommand):
 
         # Create youtube videos
         CHARSET = string.ascii_letters + string.digits
-        for i in range(100):
+        for _ in range(100):
             video_code = "".join(random.choices(CHARSET, k=7))
             YoutubeLink(url=f"https://www.youtube.com/watch?v={video_code}").save()
 
         # Create muscles
-        for muscle_abbrev, muscle_name in Muscle.MUSCLES:
+        for muscle_abbrev, _ in Muscle.MUSCLES:
             Muscle(name=muscle_abbrev).save()
 
     def create_exercises(self):
@@ -145,7 +182,12 @@ class Command(BaseCommand):
                 create_routine(**routine_dict, owner=user)
 
     def create_workouts(self):
-        ...
+        users = {2: WORKOUTS_USER_1}
+
+        for user_pk, user_workouts in users.items():
+            user = User.objects.get(pk=user_pk)
+            for workout_dict in user_workouts:
+                create_workout(**workout_dict, owner=user)
 
     def handle(self, *args, **options):
         self.reset_db()
@@ -153,3 +195,4 @@ class Command(BaseCommand):
         self.create_supporting_models()
         self.create_exercises()
         self.create_routines()
+        self.create_workouts()
